refactor: create comment.py, add comment tests
